server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "argus-mcp"
version = "0.1.0"
description = "Runtime context MCP server for AI coding agents"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "Pillow>=10.0.0",
]

[project.scripts]
argus = "argus.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/argus"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]

